+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix issue with multiple name ptrs in same message
+
 ## 0.1.0 (2020-12-17)
 
 - Initial release